Remove elements + update rubocop
