+++ .rubocop.yml
@@ -25,6 +25,3 @@ Style/MixinUsage:
   Exclude:
     - 'test/dummy/bin/setup'
     - 'test/dummy/bin/update'
-
-Style/ParallelAssignment:
-  Enabled: false
